# Conflicts:
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/example/arhiking/Models/Hike_Activity.java
@@ -3,18 +3,13 @@ package com.example.arhiking.Models;
 import androidx.room.ColumnInfo;
 import androidx.room.Entity;
 import androidx.room.PrimaryKey;
-import androidx.room.TypeConverters;
-
-import com.example.arhiking.Data.AppDatabase;
-
-import org.osmdroid.util.GeoPoint;
 
 
 @Entity
 public class Hike_Activity {
 
     @PrimaryKey(autoGenerate = true)
-    public int hikeActivityId;
+    public long hikeActivityId;
 
     @ColumnInfo(name = "hike_activity_name")
     public String hikeActivityName;
